devapp: compile static files & templates into binary
Previously the templates and static files for the dashboard and
the release dashboard were loaded from disk at runtime, each time
a page was requested. This adds an overhead to end users, who need
to ensure the source files are present on the machine running the
server, and ensure that the server's working directory is set to
$GOPATH/src/golang.org/x/build/devapp.

Instead, build these files into the binary, so we don't need to have
the source code on deployment machines, or to run the binary from
a specific working directory. Add a generator so you can run "go
generate golang.org/x/build/devapp" to regenerate the static assets in
the binary.

To speed up local development, provide a --templates command line flag
that lets you specify the static directory manually, and load the
files from disk (instead of from the compiled binary), matching the
option provided in godoc.

We used to compile the release dashboard template each time we request
a page, but this is wasteful since nothing in that template changes.
Let's compute it once and use the one compiled template to serve all
requests.

Add basic tests that the handlers in question don't panic when they
get a request.

Change-Id: Ib05acd267d956e2227795d4546ae9d66cbf87c89