all: fix vet errors
Replace oauth2.NoContext (deprecated) with context.Background(),
which has been available for two consecutive releases.

Add more cloud.google.com/go packages to the cmd/coordinator
Dockerfile to fix an error building cmd/coordinator. A dependency is
not present in master, but was present in the older revision, and was
not getting checked out correctly during the "go get" step. I added
instructions for hopefully doing this more systematically in the
future.

Change-Id: Id6c2220482350a686b87742ec7915c457a689e52